dom/base/test/test_meta_viewport_maximum_scale_2.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>maximum-scale=2 in meta viewport</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <meta name="viewport" content="maximum-scale=2">
  <script src="viewport_helpers.js"></script>
</head>
<body>
  <p>maximum-scale=2</p>
  <script type="application/javascript">
    "use strict";

    add_task(async function maximum_scale_2() {
      await SpecialPowers.pushPrefEnv(scaleRatio(1.0));

      let info = getViewportInfo(800, 480);
      is(info.width, 800, "width should be the display width");
      is(info.height, 480, "height should be the display height");
      is(info.maxZoom, 2, "maximum-scale should be 2");
    });
  </script>
</body>
</html>
